here's a basic GUI I made for testing. it's pretty nice ngl and has like a login thing which is neat. Right now all it really does is list the most common words from a user's messages and filter out filler words.
I need full messages to be about to calculate correlations between different words, and then we'll probably have to manually make a few groups of words like "happy and joyful". probably gonna be easiest with polls for alot of it.

then once we establish a baseline, then we can collect dynamic averages for different circles of though and make the graph thing like a pokedex.