gui is still shit. the main goal was the pokemon skills style bias radar graph. data is skewed because I copy and pasted it like a retard because kairoz refuses to provide good data. Also the code is very messy, I apologize. There's probably a few functions that aren't in use anymore because I pivoted to empathy halfway through. The code will continue to bloat until morale improves. Once again, if you're interested in how i collected this data, It was by hand.
either way, I used a library from the jawn to use the thing and it's a basic framework based on the newyork times text patterns or something. Then copy pasted a conversation from a forum (consent was given) and that log is the simulated_room jsonl file if you wanna play with it. be warned, it's janky BECAUSE NOBODY ELSE HAS ANY SORT OF GUI SETUP FOR TESTING ANYTHING AND THERE'S NO REAL DATA. Either way it's not gonna be accurate for this group without REAL DATA.

if you're interested in helping, please spam kairoz until bro give up da booty.